# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sets/": {
      "post": {
        "tags": [
          "sets"
        ],
        "summary": "Create Set",
        "operationId": "create_set_api_sets__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Set"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "sets"
        ],
        "summary": "Read Sets",
        "operationId": "read_sets_api_sets__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "creator_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Creator Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sets/merge": {
      "post": {
        "tags": [
          "sets"
        ],
        "summary": "Merge Sets",
        "description": "Merge multiple source sets into a single target set.\n\nAll images from the source sets will be reassigned to the target set. The source sets will then be permanently deleted. Use this to consolidate duplicate or fragmented collections.",
        "operationId": "merge_sets_api_sets_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMerge"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Set"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sets/import": {
      "post": {
        "tags": [
          "sets"
        ],
        "summary": "Import Set",
        "operationId": "import_set_api_sets_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Set"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sets/batch-import": {
      "post": {
        "tags": [
          "sets"
        ],
        "summary": "Batch Import Sets",
        "description": "Unified route to scan, parse, and optionally execute batch imports of directories.\n\nIf `dry_run=True`, it only scans the provided paths, attempts to parse creator/set names using the `parsing_template`, and returns a preview of what would be imported. If `dry_run=False`, it launches the actual import process as an asynchronous background task.",
        "operationId": "batch_import_sets_api_sets_batch_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchImportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sets/{set_id}": {
      "get": {
        "tags": [
          "sets"
        ],
        "summary": "Read Set",
        "operationId": "read_set_api_sets__set_id__get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Set"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "sets"
        ],
        "summary": "Update Set",
        "operationId": "update_set_api_sets__set_id__patch",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Set Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Set"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sets"
        ],
        "summary": "Delete Set",
        "operationId": "delete_set_api_sets__set_id__delete",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Set"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sets/bulk-update": {
      "post": {
        "tags": [
          "sets"
        ],
        "summary": "Bulk Update Sets",
        "operationId": "bulk_update_sets_api_sets_bulk_update_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetBulkUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Bulk Update Sets Api Sets Bulk Update Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sets/bulk-delete": {
      "post": {
        "tags": [
          "sets"
        ],
        "summary": "Bulk Delete Sets",
        "operationId": "bulk_delete_sets_api_sets_bulk_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "integer"
                },
                "type": "array",
                "title": "Set Ids"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Bulk Delete Sets Api Sets Bulk Delete Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sets/{set_id}/resync": {
      "post": {
        "tags": [
          "sets"
        ],
        "summary": "Resync Set",
        "description": "Resynchronize a set with its local filesystem directory.\n\nScans the `local_path` associated with the set. Any new image files found in the directory that aren't already in the database will be imported and added to the set.",
        "operationId": "resync_set_api_sets__set_id__resync_post",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Set"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/images/set/{set_id}": {
      "post": {
        "tags": [
          "images"
        ],
        "summary": "Create Image For Set",
        "operationId": "create_image_for_set_api_images_set__set_id__post",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Set Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Image"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchImportRequest": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SetBatchImport"
            },
            "type": "array",
            "title": "Items",
            "description": "List of directories to import.",
            "default": []
          },
          "scan_auto_path": {
            "type": "boolean",
            "title": "Scan Auto Path",
            "description": "If true, automatically scans the configured auto-import directory.",
            "default": false
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run",
            "description": "If true, only returns what would happen without actually importing.",
            "default": true
          },
          "parsing_template": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parsing Template",
            "description": "Regex template used to extract Creator and Set names from the directory structure (e.g., '[Creator] - [Set]')."
          },
          "delete_source_default": {
            "type": "boolean",
            "title": "Delete Source Default",
            "description": "Default value for delete_source applied to scanned items.",
            "default": false
          }
        },
        "type": "object",
        "title": "BatchImportRequest"
      },
      "BatchImportResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchImportItem"
            },
            "type": "array",
            "title": "Items",
            "description": "Status of each imported item."
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary",
            "description": "Summary statistics of the import operation.",
            "default": {}
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id",
            "description": "Background task ID if the import was queued asynchronously."
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "Overall status of the batch import request."
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "BatchImportResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Set": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "The display title of the set."
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url",
            "description": "Original URL where the set was downloaded from."
          },
          "local_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Path",
            "description": "Local filesystem path where the set's files reside."
          },
          "phash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phash",
            "description": "Perceptual hash representative of the set (usually the cover image)."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "User-provided notes or context for the set."
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "Comma-separated descriptive tags for the set."
          },
          "id": {
            "type": "integer",
            "title": "Id",
            "description": "Unique database identifier for the set."
          },
          "date_added": {
            "type": "string",
            "title": "Date Added",
            "description": "Timestamp when the set was added to the vault."
          },
          "creators": {
            "items": {
              "$ref": "#/components/schemas/Creator"
            },
            "type": "array",
            "title": "Creators",
            "description": "List of creators associated with this set.",
            "default": []
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/Image"
            },
            "type": "array",
            "title": "Images",
            "description": "List of images contained in this set.",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "date_added"
        ],
        "title": "Set"
      },
...
      "SetCreate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "The display title of the set."
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url",
            "description": "Original URL where the set was downloaded from."
          },
          "local_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Path",
            "description": "Local filesystem path where the set's files reside."
          },
          "phash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phash",
            "description": "Perceptual hash representative of the set (usually the cover image)."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "User-provided notes or context for the set."
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "description": "Comma-separated descriptive tags for the set."
          },
          "creator_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Creator Ids",
            "description": "List of creator IDs to associate with this set.",
            "default": []
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageCreate"
            },
            "type": "array",
            "title": "Images",
            "description": "List of images to create and associate with this set.",
            "default": []
          }
        },
        "type": "object",
        "title": "SetCreate"
      },
      "SetImport": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Title of the imported set."
          },
          "creator_names": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Creator Names",
            "description": "List of creator names extracted during import.",
            "default": []
          },
          "local_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Path",
            "description": "Path to the directory containing the set."
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/ImageCreate"
            },
            "type": "array",
            "title": "Images",
            "description": "Images found during the import process.",
            "default": []
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "description": "Auto-generated or extracted notes."
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "SetImport"
      },
      "SetMerge": {
        "properties": {
          "source_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Source Ids",
            "description": "List of set IDs to merge. These sets will be deleted."
          },
          "target_id": {
            "type": "integer",
            "title": "Target Id",
            "description": "The ID of the set that will receive all images from the source sets."
          }
        },
        "type": "object",
        "required": [
          "source_ids",
          "target_id"
        ],
        "title": "SetMerge"
      },
      "SetPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Set"
            },
            "type": "array",
            "title": "Items",
            "description": "Paginated list of sets."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of sets matching the query."
          },
          "skip": {
            "type": "integer",
            "title": "Skip",
            "description": "Number of items skipped."
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Maximum number of items returned."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "SetPage"
      },
...
    }
  }
}